Add stale branches heads to finality notifications
Warning. Previous implementation was sending a notification for
each block between the previous (explicitly) finalized block and
the new finalized one (with an hardcoded limit of 256).

Now the notification is sent only for the new finalization head with a
message containing the hash of the new finalized top and the previous
one. All the blocks between these two are implicitly finalized and it is
a subscriber duty to eventually walk between the two if some auxiliary
action is required.